Better --config parameter handling
We support a --config parameter to set the configuration file that
will be used for defaults (instead of the default conf file). Because
of that we need to parse the arguments looking for that parameter
before the OptionParser is used to parse the parameters.

Before that change we didn't support --config=FILE as it was advertised
by --help and the README.

Besides we now report to the user if he specified a config file and
we failed to read it. The default config file is still optional.

Fixes #8