Start a thread to run the scp command
We can't set a handler in the channel at ServerInterface because paramiko
runs it in a different thread than our StreamRequestHandler so there's a
synchronisation issue (ie. the handler gets to the join before the exec
command negotiation is done).

Instead we start a thread for handling the scp command. We use a forking
model for the server, but we fork *before* paramiko (and now out SCPHandler)
starts creating threads so it should be OK!